requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alembic==1.14.1
altair==5.5.0
amqp==5.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anthropic==0.52.2
anyio==4.6.2.post1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.1.0
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
bottle==0.13.2
bottle-websocket==0.2.9
cachelib==0.13.0
cachetools==5.5.2
celery==5.5.1
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.4.0
checksumdir==1.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
dominate==2.9.1
Eel==0.17.0
email_validator==2.2.0
executing==2.1.0
faiss-cpu==1.11.0
fastapi==0.115.13
fastjsonschema==2.20.0
filelock==3.16.1
filetype==1.2.0
Flask==3.1.1
Flask-Bootstrap==3.3.7.1
Flask-Cache==0.2
Flask-Caching==2.3.1
Flask-Celery-Helper==1.1.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security==5.6.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
gevent==24.10.3
gevent-websocket==0.10.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google_search_results==2.4.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
groq==0.29.0
grpcio==1.72.1
grpcio-status==1.71.0
h11==0.14.0
hf-xet==1.1.3
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.1
huggingface-hub==0.32.4
idna==3.7
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.29.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
json5==0.9.28
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kombu==5.5.2
langchain==0.3.26
langchain-anthropic==0.3.15
langchain-community==0.3.26
langchain-core==0.3.66
langchain-google-genai==2.0.10
langchain-groq==0.3.4
langchain-huggingface==0.2.0
langchain-openai==0.3.21
langchain-text-splitters==0.3.8
langsmith==0.3.45
Mako==1.3.9
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
multidict==6.5.1
mypy_extensions==1.1.0
narwhals==1.44.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.2.1
openai==1.84.0
orjson==3.10.18
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.2.0
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.21.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
selenium==4.23.1
Send2Trash==1.8.3
sentence-transformers==4.1.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.34
stack-data==0.6.3
starlette==0.46.2
streamlit==1.46.0
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.9.0
timeago==1.0.15
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
trio==0.26.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.2.2
uvicorn==0.34.3
vine==5.1.0
virtualenv==20.27.1
visitor==0.1.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
whichcraft==0.6.1
wikipedia==1.4.0
wsproto==1.2.0
WTForms==3.1.2
yarl==1.20.1
zope.event==5.0
zope.interface==7.1.1
zstandard==0.23.0
